requirements.txt
yfinance
pandas
google-generativeai
lxml
requests
html5lib